internship-qa-automation/requirements.txt
pytest==8.3.3
pytest-xdist==3.6.1
pytest-order==1.3.0
python-dotenv==1.0.1
requests==2.32.3
Faker==33.0.0
selenium==4.24.0
webdriver-manager==4.0.2